[demo] Using data protection to protect clusters
This demo focuses on the Data Protection feature. This feature uses
Velero and Rustic to backup and restore Kubernetes clusters in full on a
schedule. To prove that, this demo has you attach an external unmanaged
cluster, take a backup, blow everything in the cluster away and restore
it to last known good.

Signed-off-by: Carlos Nunez <ncarlos@vmware.com>